Add files via upload
Notebook shows the analysis of a survey data. The essence of the analysis is to get insights into the eating habits of young people in South Africa. The dataset co gained a column (level2 coded) that has a broad category of foods in it e.g cake, potatoes however, the client wanted to analysis to be done with the foods on a granular level e.g cheesecake, potato wedges etc. . The data was then structured to allow for the analysis on a granular level.